Fix tests and remove cyclic dependencies
